FSD 아키텍쳐 기반 파일 구조 조정
