+++ Web/src/main/orbit/craco.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+    webpack: {
+        alias: {
+            '@': path.resolve(__dirname, 'src'),
+        },
+    },
+};
