+++ Web/src/main/orbit/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@craco/craco": "^7.1.0",
     "@fortawesome/free-solid-svg-icons": "^6.7.1",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@react-oauth/google": "^0.12.1",
@@ -24,11 +25,11 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test"
   },
+
   "eslintConfig": {
     "extends": [
       "react-app",
@@ -51,6 +52,7 @@
   "devDependencies": {
     "css-loader": "^7.1.2",
     "postcss-loader": "^8.1.1",
+    "react-app-rewired": "^2.2.1",
     "style-loader": "^4.0.0"
   }
 }
