+++ Web/src/main/orbit/src/App.js
@@ -1,18 +1,13 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { GoogleOAuthProvider } from '@react-oauth/google';
-import AppRoutes from './routes/AppRoutes';
+import AppRoutes from './app/routes/AppRoutes';
 
 const App = () => {
-    const clientId = '871036489093-hkfl3tack8d6d8puumpas6a0qt22jvgq.apps.googleusercontent.com';
-
     return (
-        <GoogleOAuthProvider clientId={clientId}>
-            <Router>
-                <AppRoutes />
-            </Router>
-        </GoogleOAuthProvider>
+        <Router>
+            <AppRoutes />
+        </Router>
     );
 };
 
-export default App;
+export default App; // default export 추가
