+++ Web/src/main/orbit/src/app/routes/AppRoutes.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Routes, Route, Navigate } from 'react-router-dom';
+import { LoginPage } from '@/pages/login';
+import { SignupPage } from '@/pages/signup';
+
+const AppRoutes = () => {
+    return (
+        <Routes>
+            <Route path="/" element={<Navigate to="/login" />} />
+            <Route path="/login" element={<LoginPage />} />
+            <Route path="/signup" element={<SignupPage />} />
+        </Routes>
+    );
+};
+
+export default AppRoutes;
