+++ Web/src/main/orbit/src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import store from './redux/store';
+import store from './app/redux/store';
 import App from './App';
-import './index.css'; // Pretendard 포함된 CSS 파일 가져오기
 
 ReactDOM.render(
     <Provider store={store}>
