+++ Web/src/main/orbit/src/pages/login/index.js
@@ -1,0 +1,1 @@
+export { LoginModel as LoginPage } from './model';
