+++ Web/src/main/orbit/src/pages/signup/index.js
@@ -1,0 +1,1 @@
+export { SignupModel as SignupPage } from './model';
