--- Web/src/main/orbit/src/routes/AppRoutes.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { Routes, Route, Navigate } from 'react-router-dom';
-import Login from '../components/Login/Login';
-import Signup from '../components/Signup/Signup';
-import ForgotPassword from '../components/ForgotPassword/ForgotPassword';
-import Dashboard from '../components/Dashboard/Dashboard';
-
-const AppRoutes = () => {
-    return (
-        <Routes>
-            <Route path="/" element={<Navigate to="/login" />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/signup" element={<Signup />} />
-            <Route path="/forgot-password" element={<ForgotPassword />} />
-            <Route path="/dashboard" element={<Dashboard />} />
-        </Routes>
-    );
-};
-
-export default AppRoutes;
