--- Web/src/main/orbit/src/styles/Common.module.css
@@ -1,47 +1,0 @@
-/* 공통 버튼 스타일 */
-.button {
-    padding: 10px 20px;
-    font-size: 16px;
-    border-radius: 5px;
-    border: none;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.button-primary {
-    background-color: #007bff;
-    color: white;
-}
-
-.button-primary:hover {
-    background-color: #0056b3;
-}
-
-.button-secondary {
-    background-color: #6c757d;
-    color: white;
-}
-
-.button-secondary:hover {
-    background-color: #5a6268;
-}
-
-/* 공통 입력 필드 스타일 */
-.input {
-    width: 100%;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    margin-bottom: 10px;
-}
-
-/* 공통 링크 스타일 */
-.link {
-    color: #007bff;
-    text-decoration: none;
-}
-
-.link:hover {
-    text-decoration: underline;
-}
