--- Web/src/main/orbit/src/styles/Global.module.css
@@ -1,54 +1,0 @@
-/* 전역 초기화 */
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-
-}
-
-html, body {
-    font-family: 'Pretendard', sans-serif;
-    font-size: 16px;
-    color: #333;
-    background-color: #f8f9fa;
-    height: 100%;
-    overflow-x: hidden;
-    scroll-behavior: smooth;
-}
-
-/* 전역 컨테이너 */
-.container {
-    width: 90%;
-    max-width: 1200px;
-    margin: 0 auto;
-}
-
-/* 헤더 및 푸터 공통 스타일 */
-.header, .footer {
-    background-color: #343a40;
-    color: white;
-    text-align: center;
-    padding: 10px 0;
-}
-
-.header a, .footer a {
-    color: #007bff;
-    text-decoration: none;
-}
-
-.header a:hover, .footer a:hover {
-    text-decoration: underline;
-}
-
-/* Flex 유틸리티 */
-.flex-center {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.flex-column {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
